requirements.txt
flask
onvif-zeep
